completando clases segun diagramas

+++ entorno/src/main/java/com/app/factura/FacturaRepository.java
@@ -2,4 +2,4 @@ package com.app.factura;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FacturaRepository extends JpaRepository<Factura,Long>{}
+public interface FacturaRepository extends JpaRepository<Factura,Integer>{}
